refactor: extract the assist stream registry into its own module
Pure behavior-preserving move to keep stream.rs under the R8 hard LOC cap
after the hardening additions (stream.rs 1483 -> 896; new assist_registry.rs
643). The reqId state machine (StreamEntry, AssistStreamRegistry, the
JobCanceller/JobStarter traits, start_and_register) and its tests move out;
a re-export keeps every existing super::stream::AssistStreamRegistry path
unchanged. Test count unchanged (259).
